backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic-settings==2.6.1
email-validator==2.1.0
httpx==0.27.2
python-multipart==0.0.9
redis==5.2.1
jinja2==3.1.4
itsdangerous==2.2.0
celery==5.4.0
python-bcb==0.3.2
yfinance==0.2.51
psycopg2-binary==2.9.9
fakeredis==2.26.2
pandas==2.2.0
requests==2.32.3
slowapi==0.1.9
correpy==0.6.0
pdfplumber==0.11.4
openpyxl==3.1.5
pdf2image==1.17.0
stripe==14.4.1
boto3==1.35.0
